+++ bundle/ng2-ui-auth.js
@@ -771,7 +771,7 @@ var Oauth2Service = (function () {
         }
         return openPopup
             .switchMap(function (oauthData) {
-            if (!options.exchangeForToken && (_this.defaults.responseType === 'token' || !_this.defaults.url)) {
+            if (!options.exchangeForToken || _this.defaults.responseType === 'token' || !_this.defaults.url) {
                 return rxjs_Observable.Observable.of(oauthData);
             }
             if (oauthData.state && oauthData.state !== _this.storage.get(stateName)) {