+++ commonjs/oauth2.service.js
@@ -50,7 +50,7 @@ var Oauth2Service = (function () {
         }
         return openPopup
             .switchMap(function (oauthData) {
-            if (!options.exchangeForToken && (_this.defaults.responseType === 'token' || !_this.defaults.url)) {
+            if (!options.exchangeForToken || _this.defaults.responseType === 'token' || !_this.defaults.url) {
                 return Observable_1.Observable.of(oauthData);
             }
             if (oauthData.state && oauthData.state !== _this.storage.get(stateName)) {